# openapi.yaml
openapi: 3.0.0
info:
  title: greensolar API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3006/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /proyecto-solar:
    get:
      tags:
        - ProyectoSolar
      summary: Listar todos los ProyectoSolars
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de ProyectoSolars
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - ProyectoSolar
      summary: Crear ProyectoSolar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: ProyectoSolar creado
  /proyecto-solar/{id}:
    get:
      tags:
        - ProyectoSolar
      summary: Obtener ProyectoSolar por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ProyectoSolar encontrado
        '404':
          description: ProyectoSolar no encontrado
    patch:
      tags:
        - ProyectoSolar
      summary: Actualizar ProyectoSolar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: ProyectoSolar actualizado
    delete:
      tags:
        - ProyectoSolar
      summary: Eliminar ProyectoSolar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ProyectoSolar eliminado
  /proyecto-solar/{id}/upload-image:
    post:
      tags:
        - ProyectoSolar
      summary: Subir imagen a ProyectoSolar
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
...
  /monitoreo-consumo:
    get:
      tags:
        - MonitoreoConsumo
      summary: Listar todos los MonitoreoConsumos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de MonitoreoConsumos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - MonitoreoConsumo
      summary: Crear MonitoreoConsumo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: MonitoreoConsumo creado
  /monitoreo-consumo/{id}:
    get:
      tags:
        - MonitoreoConsumo
      summary: Obtener MonitoreoConsumo por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MonitoreoConsumo encontrado
        '404':
          description: MonitoreoConsumo no encontrado
    patch:
      tags:
        - MonitoreoConsumo
      summary: Actualizar MonitoreoConsumo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: MonitoreoConsumo actualizado
    delete:
      tags:
        - MonitoreoConsumo
      summary: Eliminar MonitoreoConsumo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: MonitoreoConsumo eliminado
  /monitoreo-consumo/{id}/upload-image:
    post:
      tags:
        - MonitoreoConsumo
      summary: Subir imagen a MonitoreoConsumo
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida